+++ main.go
@@ -19,16 +19,15 @@ package main
 import (
 	"crypto/tls"
 	"fmt"
+	"github.com/xmidt-org/argus/webhookclient"
 	"io"
 	"net/http"
 	_ "net/http/pprof"
-	"net/url"
 	"os"
 	"os/signal"
 	"runtime"
 	"time"
 
-	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
 	"github.com/xmidt-org/webpa-common/service/servicecfg"
 
@@ -62,7 +61,7 @@ func caduceus(arguments []string) int {
 		f = pflag.NewFlagSet(applicationName, pflag.ContinueOnError)
 		v = viper.New()
 
-		logger, metricsRegistry, webPA, err = server.Initialize(applicationName, arguments, f, v, Metrics, webhook.Metrics, aws.Metrics)
+		logger, metricsRegistry, webPA, err = server.Initialize(applicationName, arguments, f, v, Metrics, aws.Metrics)
 	)
 
 	if parseErr, done := printVersion(f, arguments); done {
@@ -82,13 +81,7 @@ func caduceus(arguments []string) int {
 		return 1
 	}
 
-	var (
-		infoLog  = log.WithPrefix(logger, level.Key(), level.InfoValue())
-		errorLog = log.WithPrefix(logger, level.Key(), level.ErrorValue())
-		debugLog = log.WithPrefix(logger, level.Key(), level.DebugValue())
-	)
-
-	infoLog.Log("configurationFile", v.ConfigFileUsed())
+	logging.Info(logger).Log("configurationFile", v.ConfigFileUsed())
 
 	caduceusConfig := new(CaduceusConfig)
 	err = v.Unmarshal(caduceusConfig)
@@ -136,32 +129,29 @@ func caduceus(arguments []string) int {
 		incomingQueueDepthMetric: metricsRegistry.NewGauge(IncomingQueueDepth),
 		maxOutstanding:           0,
 	}
+	measures := NewMeasures(metricsRegistry)
 
-	webhookFactory, err := webhook.NewFactory(v)
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error creating new webhook factory: %s\n", err)
-		return 1
+	var updateListSizeMetric webhookclient.ListenerFunc
+	updateListSizeMetric = func(hooks []webhook.W) {
+		measures.WebhookListSize.Set(float64(len(hooks)))
 	}
-	webhookRegistry, webhookHandler := webhookFactory.NewRegistryAndHandler(metricsRegistry)
-	webhookFactory.SetExternalUpdate(caduceusSenderWrapper.Update)
+	webhookRegistry, err := NewRegistry(RegistryConfig{
+		Logger:      logger,
+		Listener:    caduceusSenderWrapper.Update,
+		CacheConfig: caduceusConfig.CacheConfig,
+		ArgusConfig: caduceusConfig.ArgusConfig,
+	}, updateListSizeMetric)
 
-	primaryHandler, err := NewPrimaryHandler(logger, v, serverWrapper, &webhookRegistry)
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "Validator error: %v\n", err)
 		return 1
 	}
 
-	scheme := v.GetString("scheme")
-	if len(scheme) < 1 {
-		scheme = "https"
+	primaryHandler, err := NewPrimaryHandler(logger, v, serverWrapper, webhookRegistry)
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "Validator error: %v\n", err)
+		return 1
 	}
-
-	selfURL := &url.URL{
-		Scheme: scheme,
-		Host:   v.GetString("fqdn") + v.GetString("primary.address"),
-	}
-
-	webhookFactory.Initialize(primaryHandler, selfURL, v.GetString("soa.provider"), webhookHandler, logger, metricsRegistry, nil)
 
 	_, runnable, done := webPA.Prepare(logger, nil, metricsRegistry, primaryHandler)
 
@@ -169,19 +159,6 @@ func caduceus(arguments []string) int {
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "Unable to start device manager: %s\n", err)
 		return 1
-	}
-
-	var messageKey = logging.MessageKey()
-
-	if webhookFactory != nil {
-		// wait for DNS to propagate before subscribing to SNS
-		if err = webhookFactory.DnsReady(); err == nil {
-			debugLog.Log(messageKey, "Calling webhookFactory.PrepareAndStart. Server is ready to take on subscription confirmations")
-			webhookFactory.PrepareAndStart()
-		} else {
-			errorLog.Log(messageKey, "Server was not ready within a time constraint. SNS confirmation could not happen",
-				logging.ErrorKey(), err)
-		}
 	}
 
 	//
@@ -201,23 +178,7 @@ func caduceus(arguments []string) int {
 		e.Register()
 	}
 
-	// Attempt to obtain the current listener list from current system without having to wait for listener reregistration.
-	debugLog.Log(messageKey, "Attempting to obtain current listener list from source", "source",
-		v.GetString("start.apiPath"))
-	beginObtainList := time.Now()
-	startChan := make(chan webhook.Result, 1)
-	webhookFactory.Start.GetCurrentSystemsHooks(startChan)
-	var webhookStartResults webhook.Result = <-startChan
-	if webhookStartResults.Error != nil {
-		errorLog.Log(logging.ErrorKey(), webhookStartResults.Error)
-	} else {
-		// todo: add message
-		webhookFactory.SetList(webhook.NewList(webhookStartResults.Hooks))
-		caduceusSenderWrapper.Update(webhookStartResults.Hooks)
-	}
-
-	debugLog.Log(messageKey, "Current listener retrieval.", "elapsedTime", time.Since(beginObtainList))
-	infoLog.Log(messageKey, "Caduceus is up and running!", "elapsedTime", time.Since(beginCaduceus))
+	logging.Info(logger).Log(logging.MessageKey(), "Caduceus is up and running!", "elapsedTime", time.Since(beginCaduceus))
 
 	signals := make(chan os.Signal, 10)
 	signal.Notify(signals)